+++ spec/integration/concurrent_processing/concurrent_processing_spec.rb
@@ -4,7 +4,7 @@
 
 require 'concurrent'
 
-setup_localstack
+setup_sqs
 
 queue_name = DT.queue
 create_test_queue(queue_name)