+++ spec/integration/non_retryable_exception/non_retryable_exception_spec.rb
@@ -6,7 +6,7 @@
 
 require 'timeout'
 
-setup_localstack
+setup_sqs
 
 # Reset DT state to prevent data leakage from previous tests
 DT.clear